Finished part 1 in Piet.
The actual logic was fairly easy once you got used to everything.
The challenge was mostly in designing a fun path. It is also very
easy do do one color wrong which creates a lot of problems as it
is not that easy to debug. Part 2 is looking way harder, I am not
sure I will be able to do it in Piet