Finished part 2!
It was ctually very feasible once you got a bit deeper in understanding
how rolling works. I also removed the need to end the input file
with a line break. This is nice, but breaks part1. So I need to fix it.
This second part may also work well, but the next step should be to
make it prettier!