Finished part 2 in Pascal.
Was quite reluctant to do this as I did not find a nice analytical
solution. But when i finally did it I found it rather enjoyable. I
was quite happy with my solution of how to take steps around the
spiral. Also funny that I found the sequence on OEIS ^^