Finished part 1 in x86 assembler.
Was surprisingly easy to solve the problem. Suprisingly annoying
to read a number from input. When it came to printing a number I
re used an older function. The code will have to be refactored
into part1 and part2 functions for part 2 as they both share the
reading of input.